# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gitatrading Store API Documentation",
        "description": "REST API documentation for Gitatrading Store frontend integration with Next.js",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development"
        },
        {
            "url": "https://api.gitatrading-store.com",
            "description": "Production"
        }
    ],
    "paths": {
        "/api/v1/customer/register": {
            "post": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Register customer",
                "description": "Endpoint untuk register customer frontend Next.js dan mengembalikan bearer token.",
                "operationId": "ae7f84063dd948763951671929f2dfb8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerRegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer registered successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerAuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customer/login": {
            "post": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Login customer",
                "description": "Endpoint untuk login customer frontend Next.js menggunakan email dan password.",
                "operationId": "79c92da621e2d38676c201548d0e8e8c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer logged in successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerAuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Email atau password salah.",
                                    "errors": []
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customer/forgot-password": {
            "post": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Send reset password link",
                "description": "Endpoint untuk mengirim link reset password ke email customer. Email yang tidak terdaftar tetap mendapat response generic.",
                "operationId": "cf0ee6e2c677d67249b6e1e677909592",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerForgotPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset password link accepted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerMessageResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many reset link requests.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customer/reset-password": {
            "post": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Reset customer password",
                "description": "Endpoint untuk reset password customer menggunakan token valid dari email reset password.",
                "operationId": "745a22066ecb9a3e97f6301cc44314ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid or expired reset token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Token reset password tidak valid atau sudah kedaluwarsa.",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customer/me": {
            "get": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Get authenticated customer",
                "description": "Endpoint untuk mengambil data customer yang sedang login menggunakan bearer token.",
                "operationId": "82a95ace77966c72afa2a1ab83f6dc02",
                "responses": {
                    "200": {
                        "description": "Authenticated customer retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerMeResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customer/logout": {
            "post": {
                "tags": [
                    "Customer Authentication"
                ],
                "summary": "Logout customer",
                "description": "Endpoint untuk logout customer dengan menghapus bearer token yang sedang digunakan.",
                "operationId": "75bfdb6b9af239127424fac78e9b5cf7",
                "responses": {
                    "200": {
                        "description": "Customer logged out successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerMessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cart": {
                "required": [
                    "id",
                    "cart_token",
                    "customer_id",
                    "is_guest",
                    "total_qty",
                    "total_price",
                    "items"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9d44f8ea-71d9-4c88-939a-2aafec67dc40",
                        "nullable": true
                    },
                    "cart_token": {
                        "type": "string",
                        "example": "guest-cart-token-from-nextjs",
                        "nullable": true
                    },
                    "customer_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": null,
                        "nullable": true
                    },
                    "is_guest": {
                        "type": "boolean",
                        "example": true
                    },
                    "total_qty": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 35998000
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItem"
                        }
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-03T10:20:30.000000Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Customer": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9c5f48a7-4a56-4f71-a4a5-7a872ecfd413"
                    },
                    "name": {
                        "type": "string",
                        "example": "Jane Customer"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "081234567890",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CustomerRegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Jane Customer"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "081234567890",
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123",
                        "minLength": 8
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "CustomerLoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "CustomerForgotPasswordRequest": {
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    }
                },
                "type": "object"
            },
            "CustomerResetPasswordRequest": {
                "required": [
                    "email",
                    "token",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "token": {
                        "type": "string",
                        "example": "reset-token-from-email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "newPassword123",
                        "minLength": 8
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "newPassword123"
                    }
                },
                "type": "object"
            },
            "CustomerAuthResponse": {
                "required": [
                    "success",
                    "message",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Login berhasil."
                    },
                    "data": {
                        "required": [
                            "customer",
                            "token",
                            "token_type"
                        ],
                        "properties": {
                            "customer": {
                                "$ref": "#/components/schemas/Customer"
                            },
                            "token": {
                                "type": "string",
                                "example": "plain-text-api-token"
                            },
                            "token_type": {
                                "type": "string",
                                "example": "Bearer"
                            },
                            "cart": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/Cart"
                                    }
                                ],
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "CustomerMeResponse": {
                "required": [
                    "success",
                    "message",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "required": [
                            "customer"
                        ],
                        "properties": {
                            "customer": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "CustomerMessageResponse": {
                "required": [
                    "success",
                    "message",
                    "data"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Logout berhasil."
                    },
                    "data": {
                        "type": "object",
                        "example": []
                    }
                },
                "type": "object"
            },
            "ApiErrorResponse": {
                "required": [
                    "success",
                    "message",
                    "errors"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error"
                    },
                    "errors": {
                        "type": "object",
                        "example": []
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "required": [
                    "success",
                    "message",
                    "errors"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "per_page": [
                                "The per page field must not be greater than 100."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Customer API token",
                "scheme": "bearer"
            }
        }
    },
...
}